+++ the-front-end/front-end/src/App.js
@@ -4,6 +4,7 @@ import Login from "./pages/login/login";
 import Register from "./pages/register/register";
 import { Route, Routes } from "react-router-dom";
 import FindAtm from "./pages/findatm/findatm";
+import AccountPage from "./pages/account/account";
 function App() {
   return (
     <>
@@ -14,6 +15,7 @@ function App() {
         <Route path="/login" element={<Login />} />
         <Route path="/findatm" element={<FindAtm />} />
         <Route path="/register" element={<Register />} />
+        <Route path="/account" element={<AccountPage />} />
       </Routes>
     </>
   );
